feat: fractional Display scale (100–400%), decoupled from Resolution
Users can now choose any display scale, not just 1x/2x — e.g. 2.5x when 2x text
is too small and 4x is too big. Reframes the two knobs the way a user thinks
about them: Resolution = the encoded backing = bandwidth; Display scale = how
large the host UI is drawn. They're independent — changing scale does not change
bandwidth.

Mechanism: the 0x1d mode table already carries pixel and point dims as two
independent resolutions, so any pixel:point ratio is expressible on the wire —
no integer-scale constraint. Make hidpi_scale a float end-to-end (AdvertiseDims,
rfb build_virtual_display, session.send_dynamic_resolution, app._resolve_hidpi_
request). --hidpi now accepts auto/on/off OR a number 1.0–4.0; an explicit
--advertise WxH@N (including @2.5) is honoured (previously --hidpi silently
overrode it — that's also a latent-bug fix). The browser connect form gains a
Display-scale dropdown; for a fixed Resolution it advertises logical = backing /
scale @ scale, for Auto it rides --hidpi. The browser frontend now threads the
scale through its dynamic-resize path (was hardcoded 2x) so the default
browser+Auto config honours it too.

Host-side note: the wire expresses any ratio, but whether Apple's screensharingd
renders a non-2x backing scale is to be verified on real hardware.

+++ src/isharescreen/proxy/protocol/negotiation.py
@@ -77,7 +77,10 @@ class AdvertiseDims:
     """
     width: int = 1920
     height: int = 1080
-    hidpi_scale: int = 2   # 2 = Retina (backing 2× points); 1 = flat (backing == points)
+    # backing:point ratio. 2.0 = Retina (backing 2× points); 1.0 = flat. May be
+    # fractional (e.g. 2.5) — the mode table carries pixel & point dims
+    # independently, so any ratio is expressible on the wire.
+    hidpi_scale: float = 2.0
     width_mm: float = 300.0
     height_mm: float = 168.75  # proportional to 1920×1080
